% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.snowball_operations.R
\name{cancel_job}
\alias{cancel_job}
\title{Cancels the specified job}
\usage{
cancel_job(JobId)
}
\arguments{
\item{JobId}{The 39-character job ID for the job that you want to cancel, for example \code{JID123e4567-e89b-12d3-a456-426655440000}.}
}
\description{
Cancels the specified job. You can only cancel a job before its \code{JobState} value changes to \code{PreparingAppliance}. Requesting the \code{ListJobs} or \code{DescribeJob} action returns a job's \code{JobState} as part of the response element data returned.
}
\examples{
# This operation cancels a job. You can only cancel a job before its
# JobState value changes to PreparingAppliance.
cancel_job(
  JobId = "JID123e4567-e89b-12d3-a456-426655440000"
)

}
