% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.snowball_operations.R
\name{create_cluster}
\alias{create_cluster}
\title{Creates an empty cluster}
\usage{
create_cluster(JobType, Resources, AddressId, RoleARN, ShippingOption,
  Description = NULL, KmsKeyARN = NULL, SnowballType = NULL,
  Notification = NULL, ForwardingAddressId = NULL)
}
\arguments{
\item{JobType}{The type of job for this cluster. Currently, the only job type supported for clusters is \code{LOCAL_USE}.}

\item{Resources}{The resources associated with the cluster job. These resources include Amazon S3 buckets and optional AWS Lambda functions written in the Python language.}

\item{AddressId}{The ID for the address that you want the cluster shipped to.}

\item{RoleARN}{The \code{RoleARN} that you want to associate with this cluster. \code{RoleArn} values are created by using the \href{http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateRole.html}{CreateRole} API action in AWS Identity and Access Management (IAM).}

\item{ShippingOption}{The shipping speed for each node in this cluster. This speed doesn't dictate how soon you'll get each Snowball Edge device, rather it represents how quickly each device moves to its destination while in transit. Regional shipping speeds are as follows:
\itemize{
\item In Australia, you have access to express shipping. Typically, devices shipped express are delivered in about a day.
\item In the European Union (EU), you have access to express shipping. Typically, Snowball Edges shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.
\item In India, Snowball Edges are delivered in one to seven days.
\item In the US, you have access to one-day shipping and two-day shipping.
}}

\item{Description}{An optional description of this specific cluster, for example \code{Environmental Data Cluster-01}.}

\item{KmsKeyARN}{The \code{KmsKeyARN} value that you want to associate with this cluster. \code{KmsKeyARN} values are created by using the \href{http://docs.aws.amazon.com/kms/latest/APIReference/API_CreateKey.html}{CreateKey} API action in AWS Key Management Service (AWS KMS).}

\item{SnowballType}{The type of AWS Snowball device to use for this cluster. Currently, the only supported device type for cluster jobs is \code{EDGE}.}

\item{Notification}{The Amazon Simple Notification Service (Amazon SNS) notification settings for this cluster.}

\item{ForwardingAddressId}{The forwarding address ID for a cluster. This field is not supported in most regions.}
}
\description{
Creates an empty cluster. Each cluster supports five nodes. You use the CreateJob action separately to create the jobs for each of these nodes. The cluster does not ship until these five node jobs have been created.
}
\examples{
# Creates an empty cluster. Each cluster supports five nodes. You use the
# CreateJob action separately to create the jobs for each of these nodes.
# The cluster does not ship until these five node jobs have been created.
create_cluster(
  AddressId = "ADID1234ab12-3eec-4eb3-9be6-9374c10eb51b",
  Description = "MyCluster",
  JobType = "LOCAL_USE",
  KmsKeyARN = "arn:aws:kms:us-east-1:123456789012:key/abcd1234-12ab-34cd-56ef-123456123456",
  Notification = list(
    JobStatesToNotify = list(),
    NotifyAll = FALSE
  ),
  Resources = list(
    S3Resources = list(
      list(
        BucketArn = "arn:aws:s3:::MyBucket",
        KeyRange = structure(
          list(),
          .Names = character(
            0
          )
        )
      )
    )
  ),
  RoleARN = "arn:aws:iam::123456789012:role/snowball-import-S3-role",
  ShippingOption = "SECOND_DAY",
  SnowballType = "EDGE"
)

}
