% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.snowball_operations.R
\name{describe_job}
\alias{describe_job}
\title{Returns information about a specific job including shipping information, job status, and other important metadata}
\usage{
describe_job(JobId)
}
\arguments{
\item{JobId}{[required] The automatically generated ID for a job, for example \code{JID123e4567-e89b-12d3-a456-426655440000}.}
}
\description{
Returns information about a specific job including shipping information, job status, and other important metadata.
}
\section{Accepted Parameters}{
\preformatted{describe_job(
  JobId = "string"
)
}
}

\examples{
# This operation describes a job you've created for AWS Snowball.
\donttest{describe_job(
  JobId = "JID123e4567-e89b-12d3-a456-426655440000"
)}

}
