% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.snowball_operations.R
\name{get_snowball_usage}
\alias{get_snowball_usage}
\title{Returns information about the Snowball service limit for your account, and also the number of Snowballs your account has in use}
\usage{
get_snowball_usage()
}
\description{
Returns information about the Snowball service limit for your account, and also the number of Snowballs your account has in use.
}
\details{
The default service limit for the number of Snowballs that you can have at one time is 1. If you want to increase your service limit, contact AWS Support.
}
\section{Accepted Parameters}{
\preformatted{get_snowball_usage()
}
}

\examples{
# Returns information about the Snowball service limit for your account,
# and also the number of Snowballs your account has in use.
# 
# The default
# service limit for the number of Snowballs that you can have at one time
# is 1. If you want to increase your service limit, contact AWS Support.
\donttest{get_snowball_usage()}

}
