% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.snowball_operations.R
\name{list_clusters}
\alias{list_clusters}
\title{Returns an array of \code{ClusterListEntry} objects of the specified length}
\usage{
list_clusters(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The number of \code{ClusterListEntry} objects to return.}

\item{NextToken}{HTTP requests are stateless. To identify what object comes "next" in the list of \code{ClusterListEntry} objects, you have the option of specifying \code{NextToken} as the starting point for your returned list.}
}
\description{
Returns an array of \code{ClusterListEntry} objects of the specified length. Each \code{ClusterListEntry} object contains a cluster's state, a cluster's ID, and other important status information.
}
\examples{
# Returns an array of ClusterListEntry objects of the specified length.
# Each ClusterListEntry object contains a cluster's state, a cluster's ID,
# and other important status information.
list_clusters()

}
