% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.snowball_operations.R
\name{update_job}
\alias{update_job}
\title{While a job's \code{JobState} value is \code{New}, you can update some of the information associated with a job}
\usage{
update_job(JobId, RoleARN = NULL, Notification = NULL, Resources = NULL,
  AddressId = NULL, ShippingOption = NULL, Description = NULL,
  SnowballCapacityPreference = NULL, ForwardingAddressId = NULL)
}
\arguments{
\item{JobId}{The job ID of the job that you want to update, for example \code{JID123e4567-e89b-12d3-a456-426655440000}.}

\item{RoleARN}{The new role Amazon Resource Name (ARN) that you want to associate with this job. To create a role ARN, use the \href{http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateRole.html}{CreateRole}AWS Identity and Access Management (IAM) API action.}

\item{Notification}{The new or updated Notification object.}

\item{Resources}{The updated \code{JobResource} object, or the updated JobResource object.}

\item{AddressId}{The ID of the updated Address object.}

\item{ShippingOption}{The updated shipping option value of this job's ShippingDetails object.}

\item{Description}{The updated description of this job's JobMetadata object.}

\item{SnowballCapacityPreference}{The updated \code{SnowballCapacityPreference} of this job's JobMetadata object. The 50 TB Snowballs are only available in the US regions.}

\item{ForwardingAddressId}{The updated ID for the forwarding address for a job. This field is not supported in most regions.}
}
\description{
While a job's \code{JobState} value is \code{New}, you can update some of the information associated with a job. Once the job changes to a different job state, usually within 60 minutes of the job being created, this action is no longer available.
}
\examples{
# This action allows you to update certain parameters for a job. Once the
# job changes to a different job state, usually within 60 minutes of the
# job being created, this action is no longer available.
update_job(
  AddressId = "ADID1234ab12-3eec-4eb3-9be6-9374c10eb51b",
  Description = "Upgraded to Edge, shipped to Finance Dept, and requested faster shipping speed - TS.",
  JobId = "JID123e4567-e89b-12d3-a456-426655440000",
  ShippingOption = "NEXT_DAY",
  SnowballCapacityPreference = "T100"
)

}
