% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.snowball_operations.R
\name{describe_address}
\alias{describe_address}
\title{Takes an \code{AddressId} and returns specific details about that address in the form of an \code{Address} object}
\usage{
describe_address(AddressId)
}
\arguments{
\item{AddressId}{The automatically generated ID for a specific address.}
}
\description{
Takes an \code{AddressId} and returns specific details about that address in the form of an \code{Address} object.
}
\examples{
# This operation describes an address for a job.
describe_address(
  AddressId = "ADID1234ab12-3eec-4eb3-9be6-9374c10eb51b"
)

}
